requirements.txt
torchvision
numpy
bottleneck
tensorboard
torchmetrics
rich
opencv-python
scipy
yaml
pyyaml
matplotlib
pandas
seaborn
pillow
tqdm
wandb
thop
requests
munch
ffmpeg
joblib
validators
multipledispatch
pynvml
xmltodict
setuptools
pycocotools
scikit-learn
ray